Pain and suffering

+++ csgopp/csgopp/common/vector.h
@@ -5,17 +5,17 @@ namespace csgopp::common::vector
 
 struct Vector3
 {
-    double x;
-    double y;
-    double z;
+    float x;
+    float y;
+    float z;
 
     Vector3() = default;
 };
 
 struct Vector2
 {
-    double x;
-    double y;
+    float x;
+    float y;
 
     Vector2() = default;
 };
